+++ qualitycheck-20190115/alibabacloud_qualitycheck20190115/models/_upload_data_sync_response_body.py
@@ -207,11 +207,13 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfo(DaraModel):
         self,
         condition_info: main_models.UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoConditionInfo = None,
         hit: main_models.UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHit = None,
+        llm_response: str = None,
         rid: str = None,
         tid: str = None,
     ):
         self.condition_info = condition_info
         self.hit = hit
+        self.llm_response = llm_response
         self.rid = rid
         self.tid = tid
 
@@ -232,6 +234,9 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfo(DaraModel):
         if self.hit is not None:
             result['Hit'] = self.hit.to_map()
 
+        if self.llm_response is not None:
+            result['LlmResponse'] = self.llm_response
+
         if self.rid is not None:
             result['Rid'] = self.rid
 
@@ -249,6 +254,9 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfo(DaraModel):
         if m.get('Hit') is not None:
             temp_model = main_models.UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHit()
             self.hit = temp_model.from_map(m.get('Hit'))
+
+        if m.get('LlmResponse') is not None:
+            self.llm_response = m.get('LlmResponse')
 
         if m.get('Rid') is not None:
             self.rid = m.get('Rid')
@@ -449,13 +457,17 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHitConditionHitInf
 class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHitConditionHitInfoHitKeyWordsHitKeyWord(DaraModel):
     def __init__(
         self,
+        customize_code: str = None,
         from_: int = None,
+        is_match: str = None,
         pid: int = None,
         tid: str = None,
         to: int = None,
         val: str = None,
     ):
+        self.customize_code = customize_code
         self.from_ = from_
+        self.is_match = is_match
         self.pid = pid
         self.tid = tid
         self.to = to
@@ -469,8 +481,14 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHitConditionHitInf
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.customize_code is not None:
+            result['CustomizeCode'] = self.customize_code
+
         if self.from_ is not None:
             result['From'] = self.from_
+
+        if self.is_match is not None:
+            result['IsMatch'] = self.is_match
 
         if self.pid is not None:
             result['Pid'] = self.pid
@@ -488,8 +506,14 @@ class UploadDataSyncResponseBodyDataResultInfoRulesRuleHitInfoHitConditionHitInf
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('CustomizeCode') is not None:
+            self.customize_code = m.get('CustomizeCode')
+
         if m.get('From') is not None:
             self.from_ = m.get('From')
+
+        if m.get('IsMatch') is not None:
+            self.is_match = m.get('IsMatch')
 
         if m.get('Pid') is not None:
             self.pid = m.get('Pid')